requirements.txt
Django==3.0.4
line-bot-sdk==1.8.0
gunicorn==20.0.4